AuthorJpa

+++ library/src/main/java/ru/otus/spring01/library/dao/AuthorDao.java
@@ -1,14 +1,13 @@
 package ru.otus.spring01.library.dao;
 
 import ru.otus.spring01.library.domain.Author;
-import ru.otus.spring01.library.domain.Person;
 
 import java.util.List;
 import java.util.UUID;
 
 public interface AuthorDao {
 
-    int count();
+    Long count();
 
     void insert(Author author);
